refactor mirage calls, define different parts of the mock api in different files
